refactor : Separate component business logic & event handler
